api/requirements.txt
Flask
SQLAlchemy
Flask-Login
flask_cors
discord.py
requests
requests_oauthlib
Flask-Limiter
passlib
bcrypt
